moved some files, improved authentication system (frontend)
moved some generic files in a misc folder for better organization, fixed authentication problems by using the react context. this way the states regarding the authentication are all handled by  the new AuthContextHandler, while the other pages just get the states from there. All the logic regarding authentication is entrusted to the AuthContextHandler.tsx,and that is why the Routes component inside AppRouter is now encapsulated inside the AuthProvider. Also removed the unnecessary search bar and the new badge from the navbar