(backend and frontend) fixed wrong variable name in getUserGames.php, finished filtering feature
now filters work in the backend, so when a user selects a tag, a request is sent to the backend to retrieve the games with that tag. this also works for searched text, but as it is now every time the user types a letter a new request is sent. since this is a huge waste of resources i am planning to make it so that the frontend sends a request only after some seconds of not typing anything in the search bar (also removed outline from focused inputs in the credentialsCard component)