(frontend) worked on the game page. added various components, made GameInfoCard more reusable, removed old imgurl field for games and replaced it with 2 new image url fields
improved the game page by using the previously imported carousel and by adding a card that displays a game's banner and its info. made the GameInfoCard component more modular, created a TagList component which dinamically displays tags (and those that are not visible are put inside a popover button which, when clicked, expands and shows the rest of the tags). the popover component was imported from a website and modified (so it was not made entirely by me). Also changed the DB structure by removing the old img_url from the game table and adding 2 new fields in its stead: a horizontal and a vertical banner urls