(frontend) added pc requirements tab to GamePage.tsx, finished ComponentsList.tsx, added fetching of user's pc components to App.tsx
ComponentsList.tsx now shows all components of a pc (planning to show a comparison to another pc), App.tsx now fetches the user's pc configuration and shows a warning and a button to add a configuration if the user doesn't have one yet or alternatively shows a button to edit the current configuration. Added the ability to fetch a user's pc from the backend (backend still lacks the file to retrieve so i used a hardcoded value for testing). Slighty changed Game interface to allow the fetching of pc requirements when fetching a game's details. Installed lucide react library for some cool icons.